requirements.txt
# =============================================================================
# PORTFOLIO DEPENDENCIES
# =============================================================================
# Minimal, focused dependency list. Each package serves a specific purpose.
# =============================================================================

# -----------------------------------------------------------------------------
# CORE FRAMEWORK
# -----------------------------------------------------------------------------
fastapi==0.109.0          # Modern async web framework
uvicorn[standard]==0.27.0 # ASGI server for production

# -----------------------------------------------------------------------------
# DATABASE
# -----------------------------------------------------------------------------
sqlalchemy==2.0.25        # ORM for database operations
aiosqlite==0.19.0         # Async SQLite support

# -----------------------------------------------------------------------------
# TEMPLATES & FORMS
# -----------------------------------------------------------------------------
jinja2==3.1.3             # HTML templating engine
python-multipart==0.0.6   # Form data parsing

# -----------------------------------------------------------------------------
# AUTHENTICATION
# -----------------------------------------------------------------------------
python-jose[cryptography]==3.3.0  # JWT token handling
passlib[bcrypt]==1.7.4    # Password hashing
bcrypt==4.1.2             # Bcrypt algorithm

# -----------------------------------------------------------------------------
# CONTENT
# -----------------------------------------------------------------------------
markdown==3.5.2           # Markdown to HTML conversion
pygments==2.17.2          # Syntax highlighting for code blocks
python-slugify==8.0.1     # URL-friendly slugs

# -----------------------------------------------------------------------------
# CONFIGURATION
# -----------------------------------------------------------------------------
python-dotenv==1.0.0      # Environment variable management
pydantic-settings==2.1.0  # Settings with validation

# -----------------------------------------------------------------------------
# DEVELOPMENT (Optional - Comment out for production)
# -----------------------------------------------------------------------------
# pytest==8.0.0           # Testing framework
# httpx==0.26.0           # Async HTTP client for tests
